Factor out FreePathFinding
